move business profile page into component folder
