+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
     "plugins":[
         "react",
-        "eslint-plugin-jsx-a11y",
+        "jsx-a11y",
         "import"
     ],
     "extends": "airbnb"
